serialVisualization/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.17.0
